reactor: crate type is changed to lib and crate name is changed to notify
