+++ crates/cli/Cargo.toml
@@ -17,6 +17,7 @@ path = "src/main.rs"
 utils = { path = "../utils" }
 nixlog ={ path = "../nixlog"}
 daemon ={ path = "../daemon"}
+notify = { path = "../notify" }
 
 # Third party dependencies
 clap = { version = "4.5.28", features = ["derive", "cargo"] }
