+++ crates/notify/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "alert"
+name = "notify"
 authors.workspace = true
 edition.workspace = true
 version.workspace = true
@@ -7,9 +7,9 @@ license.workspace = true
 homepage.workspace = true
 repository.workspace = true
 
-[[bin]]
-name = "alert"
-path = "src/main.rs"
+[lib]
+name = "notify"
+path = "src/lib.rs"
 
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
@@ -18,5 +18,3 @@ notify-rust = "4"
 relm4 = {version = "0.9.1", features=["macros", "libadwaita", "gnome_47"]}
 relm4-components = "0.9.1"
 reqwest = "0.13.2"
-tokio = {version = "1.50.0", features = ["full"]}
-zbus = "5.14.0"
