+++ crates/notify/src/lib.rs
@@ -1,0 +1,8 @@
+mod window;
+
+use std::{env, fs};
+
+pub fn modal(error: String) {
+    println!("{}", error);
+    window::open(error);
+}
